update colvars to version 2015-04-03

+++ lib/colvars/colvarbias_restraint.h
@@ -110,7 +110,7 @@ protected:
 
   /// \brief Number of steps required to reach the target force constant
   /// or restraint centers
-  size_t target_nsteps;
+  long target_nsteps;
 };
 
 /// \brief Harmonic bias restraint
